refactored config out of ml_core, renamed projects file to fpl_ml_projects
