+++ fpl_ml_projects/setup.py
@@ -1,7 +1,7 @@
 from setuptools import find_packages, setup
 
 setup(
-    name="fpl_ml",
+    name="fpl_ml_projects",
     packages=find_packages(),
     version=0.1,
     install_requires=[],
